Move WeightedTokensQueryBuilder to the core plugin
